Passport authentication added
Has been tested via Postman. Ask me for access to account that has admin access.

Added passport with jwt-strategy since we're using jwt to generate session tokens used for login. Passport finds the user information and role from the payload stored inside the tokens, and if the id checks out, we know who is trying to communicate with the server.

Authorization middleware is also added, and used to authenticate admin only requests on the user routes. At least one admin user must be added to the database manually, because change roles route requires admin role. The rest can be added via the requested route, using the access token generated by logging in as the admin that was manually created.

User model has been updated to include the fields "role" and "roleChangedAt", this is because the Discgolf app needs to show different information to different users, and also to only allow admins and clubowners to change certain data.

In addition to adding passport, return information has been stripped down to avoid unnecessary information, including email that could be PII (Personally Identifiable Information) unless necessary (like when you're getting the user by ID or Email, which should ideally only be available to those with the correct role)