Club page routes
Unfinished version of club page routes. Needs to model the club page elements and add CRUD operations on them. This can only be done after we are done modelling the elements. Meanwhile, CRUD operations for managing the club pages themselves have been added and tested in POSTMAN with authentication and authorization.

Extra:
- Refactored code in userRoutes to look nicer.
- Removed mongoose import from userService because it's not used directly.
- Added cors (.env file must be updated, ask me for access)