Refactor Togglebox
The website should work exactly as it did before. This is a refactor that aims to make the code more readable, more maintainable and scaleable. The ToggleBox component now handles the opening and closing of the box, while the children property that is passed into it will contain the elements seen within the box when open.

FirstBox was renamed DiscgolfInfo (name can be changed, I suck at naming things)

SecondBox was renamed Navigation (since it contains those navigation boxes/cards). Name can be changed on request.

Overall, this PR does not change ANYTHING about what you see on screen, it should only make the code more readable.

Lets hope the Heroku gods are pleased with my humble offering.