First version of project report
- Includes tentative structure with titles and subtitles
- Does not include any text yet

Still unsure if we should include our own front page as we described it in mandatory assignment 2, or if we should use USN's bachelor front page template. Lets discuss!

More work can be done on the report within this sprint, we already have a bunch of text we can add from the mandatory assignments. This is just to get a temporary solution out for easy viewing. The word document is available in the discord channel for editing.