Clubpage functionaly should work now
Oh boy, this is a big one.

Summary:
Clubpage should now show elements to the correct users, members and clubowner.

Clubowner can toggle between nonmember and member elements in the toolbox on the left when open.

The toolbox starts as closed so that clubowner can get a better view of the clubpage when entering the site.

Elements are only able to be updated by club owner.

Only genuine members of the club can view memberElements. Everyone else sees nonmemberElements. Clubowner sees nonmemberElements by default, but can toggle member view to see memberElements.

PS: Adminklubben is deleted as the Clubpage model was updated, please ask me for new credentials for testing purposes.